+++ src/app/pages/lista-factura/lista-factura.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ListaFacturas } from '../lista-facturas';
 import { FacturaService } from 'src/app/service/factura.service';
+import { LoginService } from 'src/app/service/login.service';
+import { MatSnackBar } from '@angular/material/snack-bar';
 
 
 @Component({
@@ -11,7 +13,7 @@ import { FacturaService } from 'src/app/service/factura.service';
 export class ListaFacturaComponent implements OnInit {
 
   lista:ListaFacturas[];
-  constructor(private facturaService:FacturaService) { }
+  constructor(private facturaService: FacturaService, private serviciologin: LoginService, private snack:MatSnackBar) { }
 
   ngOnInit(): void {
       this.facturaService.getallfactura().subscribe(
@@ -19,11 +21,19 @@ export class ListaFacturaComponent implements OnInit {
       );
   }
 
-  delete(lista:ListaFacturas):void{
-    this.facturaService.deletefactura(lista.aIdFactura).subscribe(
-      res=>this.facturaService.getallfactura().subscribe(
-        response=>this.lista=response
+  delete(lista: ListaFacturas): void {
+    let rol = this.serviciologin.getUserRol();
+
+    if (rol == 'Administrador') {
+      this.facturaService.deletefactura(lista.aIdFactura).subscribe(
+        res => this.facturaService.getallfactura().subscribe(
+          response => this.lista = response
+        )
       )
-    )
+    }
+    else {
+      this.snack.open('No tiene el privilegio suficiente', 'Aceptar', { duration: 4000, horizontalPosition: 'right' });
+      return;
+    }
   }
 }
